requirements.txt
# requirements.txt

chromadb==0.4.22
fastapi
uvicorn
python-telegram-bot==20.6
sentence-transformers
PyPDF2
pdf2image
pytesseract
aiohttp
requests
Pillow
numpy==1.26.4
posthog==2.4.2